% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRatios.R
\name{computeRatios}
\alias{computeRatios}
\title{Compute Ratios}
\usage{
computeRatios(data, groupBy, dimensions, target, distanceFromMean = TRUE)
}
\arguments{
\item{data}{\code{data.frame} with input data}

\item{groupBy}{grouping entries by specific column name}

\item{dimensions}{vector with column names used for computing ratios}

\item{target}{column name with target numeric values to compute rarios}

\item{distanceFromMean}{boolean to compute only ratios or also distance from mean value in column}
}
\value{
data frame with ratios
}
\description{
Compute Ratios
}
\examples{
library("ratan")
r <- computeRatios(
    as.data.frame(Titanic),
    "Class",
    c("Sex", "Age", "Survived"),
    "Freq")
}
