% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize}
\usage{
visualize(df)
}
\arguments{
\item{df}{\code{data.frame} with input data}
}
\value{
plotly object
}
\description{
Visualize
}
\examples{
library("ratan")
df <- computeRatios(
    as.data.frame(Titanic),
    "Class",
    c("Sex", "Age", "Survived"),
    "Freq")
visualize(df)
}
